+++ vision/yolo/model.py
@@ -1,4 +1,4 @@
-"""This file contains the class that implements and runs the YOLOv9 on images we capture."""
+"""This file contains the class that implements and runs the YOLO model on images we capture."""
 
 import asyncio
 import logging
@@ -33,7 +33,7 @@ CLASS_NAMES: dict[int, str] = {
 }
 
 # The confidence needed to accept a prediction from the YOLO model
-CONFIDENCE_THRESHOLD: float = 0.3
+CONFIDENCE_THRESHOLD: float = 0.01
 
 
 class ImageTrigger:
@@ -208,14 +208,14 @@ class ObjectDetection:
         return self._shape
 
 
-class YOLOv9:
+class YOLO:
     """
-    This class implements and runs the YOLOv9 model on images we capture.
+    This class implements and runs YOLO models on images we capture.
 
     Parameters
     ----------
     model_path : str, default="yolov9-m-converted.onnx"
-        Path to the YOLOv9 model file, by default "yolov9-m-converted.onnx"
+        Path to the YOLO model file, by default "yolov9-m-converted.onnx"
     log_results : bool, default=False
         Whether to log the results of the object detection, by default False
 
@@ -223,7 +223,7 @@ class YOLOv9:
     ----------
     onnx_session : onnxruntime.InferenceSession
         The inference session with the needed settings for
-        inferencing the YOLOv9 model.
+        inferencing the YOLO model.
     model_output : list[onnxruntime.NodeArg]
         A list of outputs generated by the model.
     output_names : list[str]
@@ -249,7 +249,7 @@ class YOLOv9:
     _filter_output(output: list[np.ndarray], image_size: tuple[int, int])
         Converts the output array of the model to the best predictions of the model.
     process_image(image_path: str)
-        Processes the image using the YOLOv9 model and returns the best predictions.
+        Processes the image using the YOLO model and returns the best predictions.
     """
 
     def __init__(
@@ -261,7 +261,7 @@ class YOLOv9:
             raise FileNotFoundError(
                 f"Model file {full_model_path} not found. Check the documentation \
                 website for a link to download the models, \
-                and place them in the vision/yolov9/models folder."
+                and place them in the vision/yolo/models folder."
             )
 
         opt_session = onnxruntime.SessionOptions()
@@ -271,7 +271,7 @@ class YOLOv9:
         self.onnx_session: onnxruntime.InferenceSession = onnxruntime.InferenceSession(
             full_model_path, sess_options=opt_session, providers=providers
         )
-        logging.info("YOLOv9 model loaded: %s", model_path)
+        logging.info("YOLO model loaded: %s", model_path)
         self.model_output: list[onnxruntime.NodeArg] = self.onnx_session.get_outputs()
         self.output_names: list[str] = [
             self.model_output[i].name for i in range(len(self.model_output))
@@ -346,6 +346,7 @@ class YOLOv9:
         # Get just the bounding box info, transpose
         boxes: npt.NDArray[np.float32] = inferences[:4, :].T
 
+        logging.info("Classes: %s", set(classes))
         # We only care about the classes that will appear in competition, choose the best one
         best_guesses: dict[int, tuple[npt.NDArray[np.float32], float]] = {}
         for i in range(len(classes)):
